Add equals() method to Take
In WorkbookRepository, if an AssociatedAudio from the db has a selected take, duplicate takes are being created for the selected take- one ends up in the selected take relay, and the other ends up in the takes relay. Unless this is changed, we need an equals function in Take to compare these different but equivalent takes.