Move Take subscriptions after the workbook take has been created
Previously, subscriptions to the Take's 'modified' and 'deleted' timestamp were being set up when the workbook take is created in WorkbookRepository. We want to create takes elsewhere in the app as well, so we should set up these subscriptions in the Associated Audio's takesRelay.onNext{} instead. This way, it doesn't matter where the take was created.