Revised fix for duplicate workbook takes being created
We don't really need the modelToWbTakeMap, and creating it actually means that the weak references might not go away since the workbook takes are a key in one WeakHashMap and the model takes are a key in another WeakHashMap. We would have to make one of the hash maps contain a WeakReference value for this to work.