Make FloatRingBuffer general purpose
Changes the FloatRingBuffer to not have the return array be domain specific to drawing lines.

As such, the allocation of this array needs to be double the width in the active recording renderer.

Changed the compression value to reflect a bug that sort of skipped over half the array at draw time. This should still be calculated based on the screen width and desired number of seconds on the screen.